@@ -1,13 +1,15 @@
 module github.com/gnmahanth/pprof-webserver
 
-go 1.21
+go 1.24.0
+
+toolchain go1.24.10
 
 require (
-	github.com/google/pprof v0.0.0-20240125082051-42cd04596328
+	github.com/google/pprof v0.0.0-20251114195745-4902fdda35c8
 	github.com/sirupsen/logrus v1.9.3
 )
 
 require (
-	github.com/ianlancetaylor/demangle v0.0.0-20240123011522-1fd6b93ad1eb // indirect
-	golang.org/x/sys v0.16.0 // indirect
+	github.com/ianlancetaylor/demangle v0.0.0-20251118225945-96ee0021ea0f // indirect
+	golang.org/x/sys v0.38.0 // indirect
 )
